Add item definitions page and loader
