+++ index.html
@@ -31,7 +31,6 @@
   <body class="hexBackground">
     <span class="criss debug"></span>
     <span class="cross debug"></span>
-
     <div id="viewport">
       <div class="level 3x1 contain" id="level-one">
         <div class="screen">
@@ -106,36 +105,9 @@
           <div
             class="object pickup on-touch respawn"
             id="mushroom"
+            data-template="item-shroom"
             onclick="console.log('picked up mushroom')"
-          >
-            <div
-              class="item name-shroom"
-              onclick="console.log('used mushroom')"
-            >
-              <div class="iconElement shroom-icon"></div>
-              <span class="inspectElement mushroom-inspect">
-                <h2>This is the dev mushroom</h2>
-                <p>Click the button below to toggle the broadcast on or off</p>
-                <button
-                  onclick="
-                  console.log('awooga')
-                  const broadcastValue = game.signalManager.checkBroadcast('toggle_two');
-                  if (broadcastValue === 'on') {
-                    game.signalManager.broadcastSignal('toggle_two', 'off');
-                  } else {
-                    game.signalManager.broadcastSignal('toggle_two', 'on');
-                  }
-                "
-                >
-                  Toggle Broadcast
-                </button>
-              </span>
-              <span class="description">
-                A mushroom that a plumber told you would increase your size
-              </span>
-              <span class="count">1</span>
-            </div>
-          </div>
+          ></div>
         </div>
         <div class="screen">
           <div class="object-label debug">Room 2</div>
