+++ js/game.js
@@ -125,7 +125,7 @@ class Game {
         if (this.paused) {
             this.pauseElement.style.visibility = 'visible';
             this.pauseElement.style.pointerEvents = 'all';
-            this.player.inventory.HUD.syncInventoryMenu();
+            this.player.HUD.syncInventoryMenu();
         } else {
             this.pauseElement.style.visibility = 'hidden';
             this.pauseElement.style.pointerEvents = 'none';
